Move trigger to a dedicated file
