+++ airflow/providers/google/cloud/operators/dataproc.py
@@ -18,7 +18,6 @@
 #
 """This module contains Google Dataproc operators."""
 
-import asyncio
 import inspect
 import ntpath
 import os
@@ -52,7 +51,7 @@ from airflow.providers.google.cloud.links.dataproc import (
     DataprocLink,
     DataprocListLink,
 )
-from airflow.triggers.base import BaseTrigger, TriggerEvent
+from airflow.providers.google.cloud.triggers.dataproc import DataprocBaseTrigger
 from airflow.utils import timezone
 
 if TYPE_CHECKING:
@@ -1000,44 +999,6 @@ class DataprocJobBaseOperator(BaseOperator):
         if self.dataproc_job_id:
             self.hook.cancel_job(project_id=self.project_id,
                                  job_id=self.dataproc_job_id, region=self.region)
-
-
-class DataprocBaseTrigger(BaseTrigger):
-    """
-    TODO: description
-    """
-
-    def __init__(self, job_id, project_id, region, gcp_conn_id, impersonation_chain):
-        super().__init__()
-        self.gcp_conn_id = gcp_conn_id
-        self.impersonation_chain = impersonation_chain
-        self.job_id = job_id
-        self.project_id = project_id
-        self.region = region
-        self.hook = DataprocHook(
-            gcp_conn_id=self.gcp_conn_id, impersonation_chain=self.impersonation_chain, async_client=True)
-
-    def serialize(self):
-        return ("airflow.providers.google.cloud.operators.dataproc.DataprocBaseTrigger",
-                {"job_id": self.job_id, "project_id": self.project_id, "region": self.region})
-
-    def get_job(self) -> Job:
-        return self.hook.get_job(
-            job_id=self.job_id,
-            project_id=self.project_id,
-            region=self.region
-        )
-
-    def job_finished(self) -> bool:
-        job = self.get_job()
-        state = job.status.state
-        return state in (JobStatus.State.ERROR, JobStatus.State.DONE, JobStatus.State.CANCELLED)
-
-    async def run(self):
-        while self.job_finished():
-            await asyncio.sleep(10)
-        yield TriggerEvent({"job": self.get_job()})
-
 
 class DataprocSubmitPigJobOperator(DataprocJobBaseOperator):
     """
